refactor: reorganize DiscordNotify class and remove unused CommandExecutor class
